Windows, 64-bit: Suppress compiler warnings about casts from SOCKET to int
Although the SOCKET data type is 64 bits wide when compiling in x64 mode,
casting from SOCKET to int is safe:
* The number of handles per process is below 2**24.
* The handle returned by the Windows kernel to a user-space process is
  an index into a table of pointers. This means that its value will remain
  relatively low even if the virtual memory page containing the handle
  object resides at an address above the 4 GB limit.
See also Mark Russinovich, Pushing the Limits of Windows: Handles, 29 Sep 2009
(http://blogs.technet.com/b/markrussinovich/archive/2009/09/29/3283844.aspx).

+++ agent/mibgroup/smux/smux.c
@@ -236,7 +236,7 @@ real_init_smux(void)
 #endif
     netsnmp_sockaddr_in( &lo_socket, smux_socket, SMUXPORT );
 
-    if ((smux_listen_sd = socket(AF_INET, SOCK_STREAM, 0)) < 0) {
+    if ((smux_listen_sd = (int) socket(AF_INET, SOCK_STREAM, 0)) < 0) {
         snmp_log_perror("[init_smux] socket failed");
         return;
     }
@@ -639,7 +639,7 @@ smux_accept(int sd)
      */
     DEBUGMSGTL(("smux", "[smux_accept] Calling accept()\n"));
     errno = 0;
-    if ((fd = accept(sd, (struct sockaddr *) &in_socket, &alen)) < 0) {
+    if ((fd = (int) accept(sd, (struct sockaddr *) &in_socket, &alen)) < 0) {
         snmp_log_perror("[smux_accept] accept failed");
         return -1;
     } else {
